// src/enums/preparations/preparationTypes.ts
export enum PREPARATION_TYPES {
  CUSTOM_PREPARATION = 'CUSTOM_PREPARATION',
  AEROPRESS = 'AEROPRESS',
  V60 = 'V60',
  CHEMEX = 'CHEMEX',
  BIALETTI = 'BIALETTI',
  PORTAFILTER = 'PORTAFILTER',
  KALITA_WAVE = 'KALITA_WAVE',
  FRENCH_PRESS = 'FRENCH_PRESS',

  TURKISH = 'TURKISH',
  AEROPRESS_INVERTED = 'AEROPRESS_INVERTED',
  DRIPPER = 'DRIPPER',
  DELTER_PRESS = 'DELTER_PRESS',
  COLD_BREW = 'COLD_BREW',
  BLUE_DRIPPER = 'BLUE_DRIPPER',
  GINA = 'GINA',
  KONO = 'KONO',
  ORIGAMI = 'ORIGAMI',
  CAFELAT = 'CAFELAT',
  OREA = 'OREA',
  COLD_DRIP = 'COLD_DRIP',
  HAND_LEVER = 'HAND_LEVER',
  FLAIR = 'FLAIR',

  APRIL_BREWER = 'APRIL_BREWER',
  ESPRO_BLOOM = 'ESPRO_BLOOM',
  FELLOW_STAGG = 'FELLOW_STAGG',
  HSIAO_50 = 'HSIAO_50',
  KARLSBADER_KANNE = 'KARLSBADER_KANNE',
  MOCCA_MASTER = 'MOCCA_MASTER',
  SIPHON = 'SIPHON',

  CAFEC_FLOWER = 'CAFEC_FLOWER',
  DECEMBER_DRIPPER = 'DECEMBER_DRIPPER',
  DECENT_ESPRESSO = 'DECENT_ESPRESSO',
  HARIO_SWITCH = 'HARIO_SWITCH',
  HARIO_WOODNECK = 'HARIO_WOODNECK',
  RATIO_SIX_COFFEE_BREWER = 'RATIO_SIX_COFFEE_BREWER',
  ROK = 'ROK',
  TORNADO_DUO = 'TORNADO_DUO',
  TRICOLATE = 'TRICOLATE',
  METICULOUS = 'METICULOUS',
}